Forgot Password Implementation
Implemented the Forgot Password Function

2 Pages Added
ForgotPassword1 - First Page that asks for Email and checks if the user exists
ForgotPassword2 - Second Page that takes in the New Password and Answer to the SQ and then performs the Password Change if Correct.

Modified Auth.py to handle new pages and proper redirects and templates.